Sample kanji from a set

Sample kanji from a set

Usage

samplekan(
  set = c("kyouiku", "jouyou", "jinmeiyou", "kanjidic"),
  size = 1,
  replace = FALSE,
  prob = NULL
)

Arguments

set

a character string specifying the set of kanjis to sample from.

size

a positive number, the number of samples.

replace

logical. Sample with replacement?

prob

currently without effect.

Value

a vector of length size containing the individual characters

Examples

(sam <- samplekan(size = 10))
#>  [1] "対" "坂" "救" "次" "死" "首" "喜" "未" "選" "任"
#> attr(,"index")
#>  [1] 253 383 788 270 276 178 755 544 477 679
lookup(sam)
#> 対 --> ON: タイ, ツイ | kun: あいて, こた.える, そろ.い, つれあ.い, なら.ぶ, むか.う | nanori: つし
#>        meaning: vis-a-vis, opposite, even, equal, versus, anti-, compare
#> 坂 --> ON: ハン | kun: さか | nanori: か, ざ
#>        meaning: slope, incline, hill
#> 救 --> ON: キュウ | kun: すく.う | nanori: 
#>        meaning: salvation, save, help, rescue, reclaim
#> 次 --> ON: ジ, シ | kun: つ.ぐ, つぎ | nanori: き, すき, つぐ, よし
#>        meaning: next, order, sequence
#> 死 --> ON: シ | kun: し.ぬ, し.に- | nanori: 
#>        meaning: death, die
#> 首 --> ON: シュ | kun: くび | nanori: おびと, こべ, す
#>        meaning: neck, counter for songs and poems
#> 喜 --> ON: キ | kun: よろこ.ぶ, よろこ.ばす | nanori: あき, きゅ, のぶ, ゆき, よし
#>        meaning: rejoice, take pleasure in
#> 未 --> ON: ミ, ビ | kun: いま.だ, ま.だ, ひつじ | nanori: 
#>        meaning: un-, not yet, hitherto, still, even now, sign of the ram, 1-3PM, eighth sign of Chinese zodiac
#> 選 --> ON: セン | kun: えら.ぶ, え.る, よ.る | nanori: 
#>        meaning: elect, select, choose, prefer
#> 任 --> ON: ニン | kun: まか.せる, まか.す | nanori: さ, とう, ひで
#>        meaning: responsibility, duty, term, entrust to, appoint