Sample kanji from a set

Sample kanji from a set

Usage

samplekan(
  set = c("kyouiku", "jouyou", "jinmeiyou", "kanjidic"),
  size = 1,
  replace = FALSE,
  prob = NULL
)

Arguments

set

a character string specifying the set of kanjis to sample from.

size

a positive number, the number of samples.

replace

logical. Sample with replacement?

prob

currently without effect.

Value

a vector of length size containing the individual characters

Examples

(sam <- samplekan(size = 10))
#>  [1] "印" "横" "誤" "想" "春" "金" "混" "朝" "暗" "他"
#> attr(,"index")
#>  [1] 566 341 979 327 201  31 780 153 387 304
lookup(sam)
#> 印 --> ON: イン | kun: しるし, -じるし, しる.す | nanori: 
#>        meaning: stamp, seal, mark, imprint, symbol, emblem, trademark, evidence, souvenir, India
#> 横 --> ON: オウ | kun: よこ | nanori: 
#>        meaning: sideways, side, horizontal, width, woof, unreasonable, perverse
#> 誤 --> ON: ゴ | kun: あやま.る, -あやま.る | nanori: 
#>        meaning: mistake, err, do wrong, mislead
#> 想 --> ON: ソウ, ソ | kun: おも.う | nanori: 
#>        meaning: concept, think, idea, thought
#> 春 --> ON: シュン | kun: はる | nanori: あずま, かす, すの, ひ, わら
#>        meaning: springtime, spring (season)
#> 金 --> ON: キン, コン, ゴン | kun: かね, かな-, -がね | nanori: かん, きむ, こ, この, ん
#>        meaning: gold
#> 混 --> ON: コン | kun: ま.じる, -ま.じり, ま.ざる, ま.ぜる, こ.む | nanori: 
#>        meaning: mix, blend, confuse
#> 朝 --> ON: チョウ | kun: あさ | nanori: あ, あそ, ささ, ちか, とも
#>        meaning: morning, dynasty, regime, epoch, period, (North) Korea
#> 暗 --> ON: アン | kun: くら.い, くら.む, くれ.る | nanori: 
#>        meaning: darkness, disappear, shade, informal, grow dark, be blinded
#> 他 --> ON: タ | kun: ほか | nanori: 
#>        meaning: other, another, the others