Sample kanji from a set

Sample kanji from a set

Usage

samplekan(
  set = c("kyouiku", "jouyou", "jinmeiyou", "kanjidic"),
  size = 1,
  replace = FALSE,
  prob = NULL
)

Arguments

set

a character string specifying the set of kanjis to sample from.

size

a positive number, the number of samples.

replace

logical. Sample with replacement?

prob

currently without effect.

Value

a vector of length size containing the individual characters

Examples

(sam <- samplekan(size = 10))
#>  [1] "泣" "皇" "居" "底" "電" "七" "問" "型" "営" "週"
#> attr(,"index")
#>  [1] 607 859 690 567 130  65 281 695 714 217
lookup(sam)
#> 泣 --> ON: キュウ | kun: な.く | nanori: 
#>        meaning: cry, weep, moan
#> 皇 --> ON: コウ, オウ | kun:  | nanori: おうじ, おお, み
#>        meaning: emperor
#> 居 --> ON: キョ, コ | kun: い.る, -い, お.る | nanori: おき, ぐ, すえ
#>        meaning: reside, to be, exist, live with
#> 底 --> ON: テイ | kun: そこ | nanori: 
#>        meaning: bottom, sole, depth, bottom price, base, kind, sort
#> 電 --> ON: デン | kun:  | nanori: 
#>        meaning: electricity
#> 七 --> ON: シチ | kun: なな, なな.つ, なの | nanori: し, しっ, な, ひち
#>        meaning: seven
#> 問 --> ON: モン | kun: と.う, と.い, とん | nanori: はる
#>        meaning: question, ask, problem
#> 型 --> ON: ケイ | kun: かた, -がた | nanori: 
#>        meaning: mould, type, model
#> 営 --> ON: エイ | kun: いとな.む, いとな.み | nanori: 
#>        meaning: occupation, camp, perform, build, conduct (business)
#> 週 --> ON: シュウ | kun:  | nanori: 
#>        meaning: week