Sample kanji from a set

Sample kanji from a set

Usage

samplekan(
  set = c("kyouiku", "jouyou", "jinmeiyou", "kanjidic"),
  size = 1,
  replace = FALSE,
  prob = NULL
)

Arguments

set

a character string specifying the set of kanjis to sample from.

size

a positive number, the number of samples.

replace

logical. Sample with replacement?

prob

currently without effect.

Value

a vector of length size containing the individual characters

Examples

(sam <- samplekan(size = 10))
#>  [1] "治" "別" "合" "音" "厳" "鉱" "冷" "章" "液" "草"
#> attr(,"index")
#>  [1] 467 459  89  40 934 831 549 391 807  68
lookup(sam)
#> 治 --> ON: ジ, チ | kun: おさ.める, おさ.まる, なお.る, なお.す | nanori: し, ぢ, はり, はる, みち
#>        meaning: reign, be at peace, calm down, subdue, quell, govt, cure, heal, rule, conserve
#> 別 --> ON: ベツ | kun: わか.れる, わ.ける | nanori: べっ
#>        meaning: separate, branch off, diverge, fork, another, extra, specially
#> 合 --> ON: ゴウ, ガッ, カッ | kun: あ.う, -あ.う, あ.い, あい-, -あ.い, -あい, あ.わす, あ.わせる, -あ.わせる | nanori: あう, あん, い, か, こう, ごお, に, ね, や, わい
#>        meaning: fit, suit, join, 0.1
#> 音 --> ON: オン, イン, -ノン | kun: おと, ね | nanori: お, と
#>        meaning: sound, noise
#> 厳 --> ON: ゲン, ゴン | kun: おごそ.か, きび.しい, いか.めしい, いつくし | nanori: いつ, いづ, きゅうら, とし
#>        meaning: stern, strictness, severity, rigidity
#> 鉱 --> ON: コウ | kun: あらがね | nanori: 
#>        meaning: mineral, ore
#> 冷 --> ON: レイ | kun: つめ.たい, ひ.える, ひ.や, ひ.ややか, ひ.やす, ひ.やかす, さ.める, さ.ます | nanori: 
#>        meaning: cool, cold (beer, person), chill
#> 章 --> ON: ショウ | kun:  | nanori: あき, あきら, あや, ふみ
#>        meaning: badge, chapter, composition, poem, design
#> 液 --> ON: エキ | kun:  | nanori: 
#>        meaning: fluid, liquid, juice, sap, secretion
#> 草 --> ON: ソウ | kun: くさ, くさ-, -ぐさ | nanori: そ, や
#>        meaning: grass, weeds, herbs, pasture, write, draft