Sample kanji from a set

Sample kanji from a set

Usage

samplekan(
  set = c("kyouiku", "jouyou", "jinmeiyou", "kanjidic"),
  size = 1,
  replace = FALSE,
  prob = NULL
)

Arguments

set

a character string specifying the set of kanjis to sample from.

size

a positive number, the number of samples.

replace

logical. Sample with replacement?

prob

currently without effect.

Value

a vector of length size containing the individual characters

Examples

(sam <- samplekan(size = 10))
#>  [1] "感" "非" "舌" "温" "示" "息" "酸" "帳" "池" "防"
#> attr(,"index")
#>  [1] 287 717 993 397 684 345 813 430 223 735
lookup(sam)
#> 感 --> ON: カン | kun:  | nanori: 
#>        meaning: emotion, feeling, sensation
#> 非 --> ON: ヒ | kun: あら.ず | nanori: 
#>        meaning: un-, mistake, negative, injustice, non-
#> 舌 --> ON: ゼツ | kun: した | nanori: 
#>        meaning: tongue, reed, clapper
#> 温 --> ON: オン | kun: あたた.か, あたた.かい, あたた.まる, あたた.める, ぬく | nanori: あつ, あつし, なお, はる, ゆ, ゆたか
#>        meaning: warm
#> 示 --> ON: ジ, シ | kun: しめ.す | nanori: 
#>        meaning: show, indicate, point out, express, display
#> 息 --> ON: ソク | kun: いき | nanori: 
#>        meaning: breath, respiration, son, interest (on money), nuture, rest, coming to an end
#> 酸 --> ON: サン | kun: す.い | nanori: 
#>        meaning: acid, bitterness, sour, tart
#> 帳 --> ON: チョウ | kun: とばり | nanori: 
#>        meaning: notebook, account book, album, curtain, veil, net, tent
#> 池 --> ON: チ | kun: いけ | nanori: 
#>        meaning: pond, cistern, pool, reservoir
#> 防 --> ON: ボウ | kun: ふせ.ぐ | nanori: あた, う, ほう
#>        meaning: ward off, defend, protect, resist